fixed calculations bugs

+++ RatingCalculator/RcLibrary/Models/DungeonMetrics.cs
@@ -7,10 +7,10 @@ using System.Threading.Tasks;
 namespace RcLibrary.Models
 {
     public class DungeonMetrics
-    {        
+    {
         public int Level { get; set; }
-        public int Base { get; set; }
-        public int Min { get { return Base - 10; } }
-        public int Max { get { return Base + 5; } }
+        public double Base { get; set; }
+        public double Min { get { return Base - 10; } }
+        public double Max { get { return Base + 5; } }
     }
 }
